ui(refactor): split superdesk-archive module
