+++ scripts/superdesk-archive/controllers/index.js
@@ -1,0 +1,3 @@
+export { BaseListController } from './BaseListController';
+export { ArchiveListController } from './ArchiveListController';
+export { UploadController } from './UploadController';
