+++ scripts/superdesk-archive/directives/GridLayout.js
@@ -1,0 +1,13 @@
+export function GridLayout() {
+    return {
+        templateUrl: 'scripts/superdesk-items-common/views/grid-layout.html',
+        scope: {items: '='},
+        link: function(scope, elem, attrs) {
+            scope.view = 'mgrid';
+
+            scope.preview = function(item) {
+                scope.previewItem = item;
+            };
+        }
+    };
+}
