+++ scripts/superdesk-archive/directives/InlineMeta.js
@@ -1,0 +1,11 @@
+export function InlineMeta() {
+    return {
+        templateUrl: 'scripts/superdesk-archive/views/inline-meta.html',
+        scope: {
+            'placeholder': '@',
+            'showmeta': '=',
+            'item': '=',
+            'setmeta': '&'
+        }
+    };
+}
