+++ scripts/superdesk-archive/directives/ItemState.js
@@ -1,0 +1,9 @@
+export function ItemState() {
+    return {
+        templateUrl: 'scripts/superdesk-archive/views/item-state.html',
+        scope: {
+            'state': '=',
+            'embargo': '='
+        }
+    };
+}
