+++ scripts/superdesk-archive/directives/ProviderMenu.js
@@ -1,0 +1,24 @@
+ProviderMenu.$inject = ['$location'];
+
+export function ProviderMenu($location) {
+    return {
+        scope: {items: '='},
+        templateUrl: 'scripts/superdesk-archive/views/provider-menu.html',
+        link: function(scope, element, attrs) {
+
+            scope.setProvider = function(provider) {
+                scope.selected = provider;
+                $location.search('provider', scope.selected);
+            };
+
+            scope.$watchCollection(function() {
+                return $location.search();
+            }, function(search) {
+                if (search.provider) {
+                    scope.selected = search.provider;
+                }
+            });
+
+        }
+    };
+}
