+++ scripts/superdesk-archive/directives/SingleItem.js
@@ -1,0 +1,10 @@
+export function SingleItem() {
+    return {
+        templateUrl: 'scripts/superdesk-archive/views/single-item-preview.html',
+        scope: {
+            item: '=',
+            contents: '=',
+            setitem: '&'
+        }
+    };
+}
