+++ scripts/superdesk-archive/services/index.js
@@ -1,0 +1,5 @@
+export { DragItemService } from './DragItemService';
+export { FamilyService } from './FamilyService';
+export { MultiService } from './MultiService';
+export { SpikeService } from './SpikeService';
+export { ArchiveService } from './ArchiveService';
