+++ scripts/superdesk-archive/tests/upload.spec.js
